13_streamlit/04_word_cloud_app/requirements.txt
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.3
plotly==5.19.0
PyPDF2==3.0.1
python-docx==0.8.11
streamlit==1.31.1
wordcloud==1.9.3